Move read_egg_csv() from records.py to utils.py
