+++ taxcalc/utils.py
@@ -1317,3 +1317,20 @@ def ce_aftertax_income(calc1, calc2,
     cedict['ceeu2'] = ce2
     # ... return cedict
     return cedict
+
+
+def read_egg_csv(vname, fpath, **kwargs):
+    """
+    Read csv file with fpath containing vname data from EGG and
+    return dict of vname data.
+    """
+    try:
+        # grab vname data from EGG distribution
+        path_in_egg = os.path.join('taxcalc', fpath)
+        vname_fname = resource_stream(
+            Requirement.parse('taxcalc'), path_in_egg)
+        vname_dict = pd.read_csv(vname_fname, **kwargs)
+    except (DistributionNotFound, IOError):
+        msg = 'could not read {} file from EGG'
+        raise ValueError(msg.format(vname))
+    return vname_dict
